03_Streamlit_prueba/prueba/requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.8.0
numpy==1.26.4
openpyxl==3.1.2
packaging==26.0
pandas==3.0.0
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.2
setuptools==82.0.0
six==1.17.0
urllib3==2.6.3
wheel==0.46.3
